refactor(server): move config allowlist checks into a verify/startup pre-pass
The deployment-verify and toml-aggregate-and-fix work threaded a
`ConfigWarnings(Arc<Mutex<..>>)` handle plus a family of collect/report
helpers through the shared `ServerVerified::new` / `fetch_and_verify_all`
pipeline, so the aggregation (and its needless mutex, source-location index,
and warning dedup) ran on every normal `server run` and every runtime submit,
not just `server verify`.

Move the outbound-HTTP allowlist checks (unregistered secrets, unauthorized
secret replacements, uncovered destinations) into `config_prepass::preflight`,
called explicitly at startup, verify, submit, and switch. The pre-pass alone
decides continue / bail / fix; the resolvers stop making that call:
`resolve_named_secrets` is now infallible and drops names the registry does not
know, and `resolve_allowed_hosts` emits its per-entry advisories with plain
`warn!`. Per-request enforcement in `http_request_policy` is unchanged and still
fails closed, so nothing is weakened. Exec-activity `secrets` (previously only
enforced by the resolver bail) are now covered by the pre-pass too.

Removes `ConfigWarnings`, `index_allowed_hosts`, the fingerprint dedup, and the
`config_warnings` field/threading/report sites. Net -265 lines.
