+++ src/command/server.rs
@@ -35,7 +35,6 @@ use crate::config::toml::ComponentLocationFetchExt as _;
 use crate::config::toml::ComponentLocationToml;
 use crate::config::toml::ComponentStdOutputToml;
 use crate::config::toml::ConfigName;
-use crate::config::toml::ConfigWarnings;
 use crate::config::toml::DatabaseConfigToml;
 use crate::config::toml::DeploymentResolved;
 use crate::config::toml::InflightSemaphoreExt as _;
@@ -59,7 +58,7 @@ use crate::config::toml::webhook::WebhookRoute;
 use crate::config::toml::webhook::WebhookRouteVerified;
 use crate::config::toml::webhook::WebhookWasmComponentConfigResolvedExt as _;
 use crate::config::toml::webhook::WebhookWasmComponentConfigVerified;
-use crate::config::toml::{AllowedHostToml, MethodsInput, MethodsInputStar, ReplaceIn};
+use crate::config::toml::{AllowedHostToml, MethodsInput, MethodsInputStar};
 use crate::config::wasm_cache_metadata_dir;
 use crate::init;
 use crate::init::Guard;
@@ -162,7 +161,7 @@ use wasm_workers::engines::EngineConfig;
 use wasm_workers::engines::Engines;
 use wasm_workers::engines::PoolingConfig;
 use wasm_workers::epoch_ticker::EpochTicker;
-use wasm_workers::http_request_policy::{AllowedHostConfig, GlobalHttpConfig, ReplacementLocation};
+use wasm_workers::http_request_policy::{AllowedHostConfig, GlobalHttpConfig};
 use wasm_workers::log_db_forwarder;
 use wasm_workers::registry::ComponentConfig;
 use wasm_workers::registry::ComponentConfigImportable;
@@ -635,184 +634,6 @@ impl RuntimeConfigAvailability {
     }
 }
 
-fn replacement_target(replace_in: ReplaceIn) -> ReplacementLocation {
-    match replace_in {
-        ReplaceIn::Headers => ReplacementLocation::Headers,
-        ReplaceIn::Body => ReplacementLocation::Body,
-        ReplaceIn::Params => ReplacementLocation::Params,
-    }
-}
-
-fn replacement_target_name(replace_in: ReplaceIn) -> &'static str {
-    match replace_in {
-        ReplaceIn::Headers => "headers",
-        ReplaceIn::Body => "body",
-        ReplaceIn::Params => "params",
-    }
-}
-
-fn allowed_host_snippet(
-    section: &str,
-    entry: &AllowedHostToml,
-    secret: &str,
-    replace_in: ReplaceIn,
-) -> String {
-    let mut entry = entry.clone();
-    entry.secrets = vec![secret.to_string()];
-    entry.replace_in = vec![replace_in];
-    let body = toml::to_string(&entry).expect("AllowedHostToml must serialize to TOML");
-    format!("[[{section}.allowed_host]]\n{body}")
-}
-
-/// A secret replacement a component requests that the operator global allowlist
-/// does not authorize. Collected across the whole deployment so every missing
-/// allowance is reported at once, letting the operator add them all in one pass.
-struct MissingSecretReplacement {
-    component_section: &'static str,
-    component_name: ConfigName,
-    entry: AllowedHostToml,
-    secret: String,
-    replace_in: ReplaceIn,
-}
-
-/// Every component's outbound HTTP `allowed_hosts`, over the same component kinds the
-/// deployment pre-pass checks. Used by `--fix` to gather referenced secret names.
-fn deployment_allowed_host_lists(deployment: &DeploymentResolved) -> Vec<&[AllowedHostToml]> {
-    let mut lists: Vec<&[AllowedHostToml]> = Vec::new();
-    lists.extend(deployment.activities_wasm.iter().map(|c| &*c.allowed_hosts));
-    lists.extend(deployment.activities_js.iter().map(|c| &*c.allowed_hosts));
-    lists.extend(deployment.webhooks_wasm.iter().map(|c| &*c.allowed_hosts));
-    lists.extend(deployment.webhooks_js.iter().map(|c| &*c.allowed_hosts));
-    lists
-}
-
-/// The `(secret name, replacement target)` pairs the operator global allowlist authorizes.
-fn global_secret_replacements(
-    global_http_config: &GlobalHttpConfig,
-) -> hashbrown::HashSet<(&str, ReplacementLocation)> {
-    global_http_config
-        .entries()
-        .iter()
-        .flat_map(|entry| {
-            entry.secret_env_mappings.iter().flat_map(|(name, _)| {
-                entry
-                    .replace_in
-                    .iter()
-                    .copied()
-                    .map(move |target| (name.as_str(), target))
-            })
-        })
-        .collect()
-}
-
-/// Append every secret replacement `entries` requests that `server_replacements`
-/// does not authorize onto `missing`. Registered-only secrets are checked (an
-/// unregistered secret name cannot be injected, so it is skipped here).
-fn collect_outbound_http_secret_replacements(
-    component_section: &'static str,
-    component_name: &ConfigName,
-    entries: &[AllowedHostToml],
-    server_replacements: &hashbrown::HashSet<(&str, ReplacementLocation)>,
-    secret_registry: &SecretRegistry,
-    missing: &mut Vec<MissingSecretReplacement>,
-) {
-    for entry in entries {
-        if entry.methods.is_none()
-            || matches!(&entry.methods, Some(MethodsInput::List(methods)) if methods.is_empty())
-        {
-            continue;
-        }
-        for secret in &entry.secrets {
-            if secret_registry.secret_lookup(secret).is_none() {
-                continue;
-            }
-            for replace_in in &entry.replace_in {
-                let requested_replacement = (secret.as_str(), replacement_target(*replace_in));
-                if server_replacements.contains(&requested_replacement) {
-                    continue;
-                }
-                missing.push(MissingSecretReplacement {
-                    component_section,
-                    component_name: component_name.clone(),
-                    entry: entry.clone(),
-                    secret: secret.clone(),
-                    replace_in: *replace_in,
-                });
-            }
-        }
-    }
-}
-
-/// Emit a single report covering every collected missing secret replacement, so
-/// the operator can add all the required allowlist entries to server.toml in one edit.
-/// Bails under strict availability; downgrades to a warning when unavailable
-/// runtime config is allowed (activation re-checks strictly).
-fn report_missing_outbound_http_secret_replacements(
-    missing: &[MissingSecretReplacement],
-    availability: RuntimeConfigAvailability,
-    warnings: &ConfigWarnings,
-) -> Result<(), anyhow::Error> {
-    if missing.is_empty() {
-        return Ok(());
-    }
-    use std::fmt::Write as _;
-    let mut details = String::new();
-    let mut server_snippets: Vec<String> = Vec::new();
-    let mut deployment_snippets: Vec<String> = Vec::new();
-    for m in missing {
-        let _ = writeln!(
-            details,
-            "  - component `{name}` (`[[{section}.allowed_host]]`) requests replacing secret \
-             `{secret}` in `{target}`",
-            name = m.component_name,
-            section = m.component_section,
-            secret = m.secret,
-            target = replacement_target_name(m.replace_in),
-        );
-        // Multiple components may request the same allowlist entry; list each unique
-        // snippet once so the operator does not paste duplicates.
-        let server_snippet =
-            allowed_host_snippet("outbound_http", &m.entry, &m.secret, m.replace_in);
-        if !server_snippets.contains(&server_snippet) {
-            server_snippets.push(server_snippet);
-        }
-        let deployment_snippet =
-            allowed_host_snippet(m.component_section, &m.entry, &m.secret, m.replace_in);
-        if !deployment_snippets.contains(&deployment_snippet) {
-            deployment_snippets.push(deployment_snippet);
-        }
-    }
-    let message = format!(
-        "the deployment requests {count} outbound HTTP secret replacement(s) that no server.toml \
-         `[[outbound_http.allowed_host]]` entry authorizes:\n\
-         {details}\n\
-         After review, add these allowlist entries to server.toml:\n\n\
-         {server}\n\
-         The corresponding deployment.toml entries are:\n\n\
-         {deployment}",
-        count = missing.len(),
-        server = server_snippets.join("\n"),
-        deployment = deployment_snippets.join("\n"),
-    );
-    if availability.allows_unavailable() {
-        warnings.insert(format!(
-            "{message}\nSkipping these load-time secret replacement checks because unavailable \
-             runtime configuration is allowed; activation will enforce them strictly."
-        ));
-        Ok(())
-    } else {
-        bail!("{message}");
-    }
-}
-
-/// A component destination that no operator global allowlist entry covers, collected
-/// across the deployment so every gap is reported at load time rather than one-by-one.
-struct UncoveredOutboundHost {
-    component_section: &'static str,
-    component_name: ConfigName,
-    entry: AllowedHostToml,
-}
-
 /// The global allowlist to enforce for a webhook. A `self_authorize` webhook (only
 /// the framework-injected, operator-owned web UI) is bounded by its own component
 /// allowlist rather than requiring an explicit server.toml entry; every other
@@ -827,108 +648,6 @@ fn webhook_global_http_allowlist(
     } else {
         operator_global.clone()
     }
-}
-
-/// Render a destination-only `[[<section>.allowed_host]]` snippet, dropping secret fields.
-fn host_allowlist_snippet(section: &str, entry: &AllowedHostToml) -> String {
-    #[derive(serde::Serialize)]
-    struct DestinationEntry {
-        pattern: String,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        methods: Option<MethodsInput>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        request_url_regex: Option<String>,
-    }
-    let body = toml::to_string(&DestinationEntry {
-        pattern: entry.pattern.clone(),
-        methods: entry.methods.clone(),
-        request_url_regex: entry.request_url_regex.clone(),
-    })
-    .expect("destination allowlist entry must serialize to TOML");
-    format!("[[{section}.allowed_host]]\n{body}")
-}
-
-/// Append every destination in `entries` that no global allowlist entry covers onto
-/// `uncovered`. Allow-nothing entries need no allowlist entry; resolution failures are
-/// left for component preparation to report authoritatively.
-#[expect(clippy::too_many_arguments)]
-fn collect_uncovered_outbound_http_hosts(
-    component_section: &'static str,
-    component_name: &ConfigName,
-    entries: &[AllowedHostToml],
-    global_http_config: &GlobalHttpConfig,
-    secret_registry: &SecretRegistry,
-    ignore_missing_env_vars: bool,
-    warnings: &ConfigWarnings,
-    uncovered: &mut Vec<UncoveredOutboundHost>,
-) {
-    for entry in entries {
-        if entry.methods.is_none()
-            || matches!(&entry.methods, Some(MethodsInput::List(methods)) if methods.is_empty())
-        {
-            continue;
-        }
-        // Resolve one entry at a time to keep the original TOML for the snippet.
-        let Ok(resolved) = resolve_allowed_hosts(
-            vec![entry.clone()],
-            ignore_missing_env_vars,
-            secret_registry,
-            warnings,
-        ) else {
-            continue;
-        };
-        let Some(resolved) = resolved.first() else {
-            continue;
-        };
-        if !global_http_config
-            .entries()
-            .iter()
-            .any(|allowed| allowed.covers(resolved))
-        {
-            uncovered.push(UncoveredOutboundHost {
-                component_section,
-                component_name: component_name.clone(),
-                entry: entry.clone(),
-            });
-        }
-    }
-}
-
-/// Warn once, listing the allowlist entries to add. A warning, not an error: coverage
-/// is conservative and a component may declare destinations it never calls.
-fn report_uncovered_outbound_http_hosts(
-    uncovered: &[UncoveredOutboundHost],
-    warnings: &ConfigWarnings,
-) {
-    if uncovered.is_empty() {
-        return;
-    }
-    use std::fmt::Write as _;
-    let mut details = String::new();
-    let mut snippets: Vec<String> = Vec::new();
-    for host in uncovered {
-        let _ = writeln!(
-            details,
-            "  - component `{name}` (`[[{section}.allowed_host]]`) allows `{pattern}`",
-            name = host.component_name,
-            section = host.component_section,
-            pattern = host.entry.pattern,
-        );
-        let snippet = host_allowlist_snippet("outbound_http", &host.entry);
-        if !snippets.contains(&snippet) {
-            snippets.push(snippet);
-        }
-    }
-    warnings.insert(format!(
-        "{count} outbound HTTP destination(s) the deployment allows are not covered by any \
-         server.toml `[[outbound_http.allowed_host]]` allowlist entry; requests to them will be \
-         denied at runtime:\n\
-         {details}\n\
-         After review, add these allowlist entries to server.toml:\n\n\
-         {snippets}",
-        count = uncovered.len(),
-        snippets = snippets.join("\n"),
-    ));
 }
 
 #[derive(Debug, Clone)]
@@ -999,7 +718,7 @@ pub(crate) async fn verify(
             &mut unregistered,
         );
         if let Some(deployment) = deployment_opt.as_ref() {
-            for hosts in deployment_allowed_host_lists(deployment) {
+            for hosts in config_prepass::deployment_allowed_host_lists(deployment) {
                 config_prepass::collect_unregistered_secrets(
                     hosts,
                     &secret_registry,
@@ -1081,6 +800,11 @@ pub(crate) async fn verify(
         .await?
     };
     let server_verified = Box::pin(server_verify(config, engines, secret_registry)).await?;
+    config_prepass::preflight(
+        &server_verified,
+        Some(&deployment),
+        verify_params.runtime_config_availability,
+    )?;
     let cas: Option<Arc<dyn concepts::cas::Cas>> = if let Some((pool, _)) = db_pool.as_ref() {
         Some(pool.cas_conn().await?.into())
     } else {
@@ -1268,11 +992,7 @@ pub(crate) async fn deployment_verify_config_compile_link(
         params.clone(),
         termination_watcher,
     )
-    .await;
-    if deployment_verified.is_err() {
-        server_verified.config_warnings.report();
-    }
-    let deployment_verified = deployment_verified?;
+    .await?;
     deployment_compile_link(
         server_verified,
         deployment_verified,
@@ -1291,7 +1011,6 @@ pub(crate) async fn deployment_compile_link(
     params: VerifyParams,
     termination_watcher: &mut watch::Receiver<()>,
 ) -> Result<ServerCompiledLinked, anyhow::Error> {
-    let config_warnings = server_verified.config_warnings.clone();
     let compiled_and_linked = ServerCompiledLinked::new(
         deployment_id,
         deployment_verified,
@@ -1300,9 +1019,7 @@ pub(crate) async fn deployment_compile_link(
         params.suppress_type_checking_errors,
         params.suppress_linking_errors,
     )
-    .await;
-    config_warnings.report();
-    let compiled_and_linked = compiled_and_linked?;
+    .await?;
     if compiled_and_linked.supressed_errors.is_none() {
         info!("Obelisk configuration was verified");
     } else {
@@ -1401,7 +1118,6 @@ pub(crate) async fn deployment_verify_config(
         server_verified.api_addr_if_webui_enabled.clone(),
         server_verified.secret_registry.clone(),
         server_verified.global_http_config.clone(),
-        server_verified.config_warnings.clone(),
     ))
     .await?;
     trace!("Verified deployment: {deployment_verified:#?}");
@@ -2015,6 +1731,11 @@ pub(crate) async fn run_internal(
         };
     let engines = create_engines(&config, &prepared_dirs)?;
     let server_verified = server_verify(config, engines, secret_registry).await?;
+    config_prepass::preflight(
+        &server_verified,
+        Some(&deployment_resolved),
+        RuntimeConfigAvailability::Strict,
+    )?;
     let cas: Arc<dyn concepts::cas::Cas> = db_pool.cas_conn().await?.into();
     let compiled_and_linked = Box::pin(deployment_verify_config_compile_link(
         server_verified.clone(),
@@ -2236,7 +1957,10 @@ pub(crate) struct ServerVerified {
     api_addr_if_webui_enabled: Option<String>,
     max_deployment_file_bytes: u32,
     global_http_config: GlobalHttpConfig,
-    config_warnings: ConfigWarnings,
+    /// The server's own `[[outbound_http.allowed_host]]` entries, verbatim. Kept so the
+    /// `config_prepass::preflight` can report unregistered secret names before they are
+    /// dropped by resolution.
+    server_outbound_allowed_hosts: Vec<AllowedHostToml>,
     /// Operator-owned secret registry, resolved before the runtime started. Carried
     /// here so runtime redeploys/submits resolve deployment secret references without
     /// re-reading the (already wiped) process environment.
@@ -2258,10 +1982,6 @@ impl ServerVerified {
         secret_registry: Arc<SecretRegistry>,
     ) -> Result<ServerVerified, anyhow::Error> {
         trace!("Using server toml: {config:#?}");
-        let config_warnings = ConfigWarnings::default();
-        if let Some(source_path) = &config.source_path {
-            config_warnings.index_allowed_hosts(source_path);
-        }
         let mut http_servers = config.http_servers;
         if config.webui.enabled {
             let webui_listening_addr = config.webui.listening_addr;
@@ -2293,35 +2013,18 @@ impl ServerVerified {
             .as_semaphore();
         let database_subscription_interruption = config.database.get_subscription_interruption();
         if config.allow_exec_activities == AllowExecActivities::AllowAny {
-            config_warnings.insert(
+            warn!(
                 "`allow_exec_activities = true` permits deployments to run arbitrary host \
                  programs; consider allowlisting reviewed scripts by content digest instead"
-                    .to_string(),
             );
         }
-        // Fail fast on the server's own config: unregistered `[[outbound_http.allowed_host]]`
-        // secrets are always fatal (the server cannot run with an invalid allowlist), but
-        // collect every missing name first so the operator can register them in one edit.
-        let mut unregistered_secrets = BTreeSet::new();
-        config_prepass::collect_unregistered_secrets(
-            &config.outbound_http.allowed_hosts,
-            &secret_registry,
-            &mut unregistered_secrets,
-        );
-        config_prepass::report_unregistered_secrets(
-            &unregistered_secrets,
-            "server.toml `[[outbound_http.allowed_host]]` entries",
-            RuntimeConfigAvailability::Strict,
-            &config_warnings,
-        )?;
+        // Unregistered secret names in these entries are reported by `config_prepass::preflight`
+        // (which runs before this) using `server_outbound_allowed_hosts`; here they resolve to
+        // nothing and are dropped.
+        let server_outbound_allowed_hosts = config.outbound_http.allowed_hosts.clone();
         let global_http_config = GlobalHttpConfig::from(
-            resolve_allowed_hosts(
-                config.outbound_http.allowed_hosts,
-                false,
-                &secret_registry,
-                &config_warnings,
-            )
-            .context("invalid server.toml `[[outbound_http.allowed_host]]` entry")?,
+            resolve_allowed_hosts(config.outbound_http.allowed_hosts, false, &secret_registry)
+                .context("invalid server.toml `[[outbound_http.allowed_host]]` entry")?,
         );
 
         Ok(Self {
@@ -2342,7 +2045,7 @@ impl ServerVerified {
             },
             max_deployment_file_bytes: config.max_deployment_file_bytes.0,
             global_http_config,
-            config_warnings,
+            server_outbound_allowed_hosts,
             secret_registry,
         })
     }
@@ -2887,6 +2590,12 @@ pub(crate) async fn submit_deployment(
         .map_err(anyhow::Error::from)?
         .into();
     let deployment_id = requested_deployment_id.unwrap_or_else(DeploymentId::generate);
+    config_prepass::preflight(
+        &server_verified,
+        Some(&deployment_resolved),
+        runtime_config_check.availability(),
+    )
+    .map_err(SubmitDeploymentError::Other)?;
     let compiled_linked = deployment_verify_config_compile_link(
         server_verified,
         prepared_dirs,
@@ -3188,6 +2897,12 @@ async fn prepare_switch_deployment(
 
     // Cold switch: validation (compile + link) always runs before enqueuing, so a
     // deployment queued for the next restart is known-good against the current host.
+    config_prepass::preflight(
+        &deployment_switch_manager.inner.server_verified,
+        Some(&target_deployment),
+        verify_params.runtime_config_availability,
+    )
+    .map_err(SwitchError::Other)?;
     let compiled_linked = deployment_verify_config_compile_link(
         deployment_switch_manager.inner.server_verified.clone(),
         &deployment_switch_manager.inner.prepared_dirs,
@@ -3769,83 +3484,13 @@ impl DeploymentVerified {
         api_addr_if_webui_enabled: Option<String>,
         secret_registry: Arc<SecretRegistry>,
         global_http_config: GlobalHttpConfig,
-        config_warnings: ConfigWarnings,
     ) -> Result<DeploymentVerified, anyhow::Error> {
         let ignore_missing_env_vars = runtime_config_availability.allows_unavailable();
         let mut deployment = deployment.into_resolved();
-        if let Some(source_path) = &deployment.source_path {
-            config_warnings.index_allowed_hosts(source_path);
-        }
         trace!("Using deployment toml: {deployment:#?}");
-        // Scoped so the `server_replacements` borrow of `global_http_config` ends
-        // before the config is moved into the deployment below. Each component's
-        // outbound HTTP entries are checked against the operator global allowlist for
-        // both unauthorized secret replacements and uncovered destinations.
-        let (missing_replacements, uncovered_hosts, unregistered_secrets) = {
-            let server_replacements = global_secret_replacements(&global_http_config);
-            let mut missing_replacements = Vec::new();
-            let mut uncovered_hosts = Vec::new();
-            let mut unregistered_secrets = BTreeSet::new();
-            let mut check =
-                |section: &'static str, name: &ConfigName, hosts: &[AllowedHostToml]| {
-                    collect_outbound_http_secret_replacements(
-                        section,
-                        name,
-                        hosts,
-                        &server_replacements,
-                        &secret_registry,
-                        &mut missing_replacements,
-                    );
-                    collect_uncovered_outbound_http_hosts(
-                        section,
-                        name,
-                        hosts,
-                        &global_http_config,
-                        &secret_registry,
-                        ignore_missing_env_vars,
-                        &config_warnings,
-                        &mut uncovered_hosts,
-                    );
-                    config_prepass::collect_unregistered_secrets(
-                        hosts,
-                        &secret_registry,
-                        &mut unregistered_secrets,
-                    );
-                };
-            for activity in &deployment.activities_wasm {
-                check(
-                    "activity_wasm",
-                    &activity.common.name,
-                    &activity.allowed_hosts,
-                );
-            }
-            for activity in &deployment.activities_js {
-                check("activity_js", &activity.name, &activity.allowed_hosts);
-            }
-            for webhook in &deployment.webhooks_wasm {
-                check(
-                    "webhook_endpoint_wasm",
-                    &webhook.common.name,
-                    &webhook.allowed_hosts,
-                );
-            }
-            for webhook in &deployment.webhooks_js {
-                check("webhook_endpoint_js", &webhook.name, &webhook.allowed_hosts);
-            }
-            (missing_replacements, uncovered_hosts, unregistered_secrets)
-        };
-        config_prepass::report_unregistered_secrets(
-            &unregistered_secrets,
-            "the deployment's outbound HTTP entries",
-            runtime_config_availability,
-            &config_warnings,
-        )?;
-        report_missing_outbound_http_secret_replacements(
-            &missing_replacements,
-            runtime_config_availability,
-            &config_warnings,
-        )?;
-        report_uncovered_outbound_http_hosts(&uncovered_hosts, &config_warnings);
+        // The outbound-HTTP allowlist pre-pass (unregistered secrets, uncovered hosts,
+        // unauthorized secret replacements) runs in `config_prepass::preflight` before this
+        // point; the authoritative per-request enforcement lives in `http_request_policy`.
         // Check uniqueness of http_server names.
         if http_servers.len()
             > http_servers
@@ -3946,7 +3591,6 @@ impl DeploymentVerified {
                 let global_executor_instance_limiter = global_executor_instance_limiter.clone();
                 let secret_registry = secret_registry.clone();
                 let global_http_config = global_http_config.clone();
-                let config_warnings = config_warnings.clone();
                 tokio::spawn(
                     async move {
                         activity_wasm
@@ -3955,7 +3599,6 @@ impl DeploymentVerified {
                                 metadata_dir,
                                 ignore_missing_env_vars,
                                 &secret_registry,
-                                config_warnings,
                                 global_http_config,
                                 global_executor_instance_limiter,
                                 fuel,
@@ -4014,7 +3657,6 @@ impl DeploymentVerified {
                 let wasm_cache_dir = wasm_cache_dir.clone();
                 let metadata_dir = metadata_dir.clone();
                 let secret_registry = secret_registry.clone();
-                let config_warnings = config_warnings.clone();
                 tokio::spawn(
                     async move {
                         webhook
@@ -4023,7 +3665,6 @@ impl DeploymentVerified {
                                 metadata_dir,
                                 ignore_missing_env_vars,
                                 &secret_registry,
-                                config_warnings,
                                 subscription_interruption,
                             )
                             .await
@@ -4122,7 +3763,6 @@ impl DeploymentVerified {
                                 wasm_cache_dir.clone(),
                                 ignore_missing_env_vars,
                                 &secret_registry,
-                                config_warnings.clone(),
                                 global_http_config.clone(),
                                 global_executor_instance_limiter.clone(),
                                 fuel,
@@ -4164,7 +3804,6 @@ impl DeploymentVerified {
                             wasm_cache_dir.clone(),
                             ignore_missing_env_vars,
                             &secret_registry,
-                            config_warnings.clone(),
                         ).await?;
                         webhooks_js_by_names.insert(k, v);
                     }
@@ -4178,7 +3817,6 @@ impl DeploymentVerified {
                             resolved_program,
                             ignore_missing_env_vars,
                             &secret_registry,
-                            &config_warnings,
                             global_executor_instance_limiter.clone(),
                         )?
                     );
@@ -5500,11 +5138,14 @@ mod tests {
     use crate::{
         command::server::{
             DeploymentRunnable, DeploymentVerified, PrepareDirsParams, RuntimeConfigAvailability,
-            ServerCompiledLinked, ServerVerified, VerifyParams,
-            collect_outbound_http_secret_replacements, collect_uncovered_outbound_http_hosts,
-            compile_activity_inline, compute_content_digest, create_engines,
-            deployment_verify_config, fix_server_exec_digests, global_secret_replacements,
-            host_allowlist_snippet, prepare_dirs, report_missing_outbound_http_secret_replacements,
+            ServerCompiledLinked, ServerVerified, VerifyParams, compile_activity_inline,
+            compute_content_digest,
+            config_prepass::{
+                collect_outbound_http_secret_replacements, collect_uncovered_outbound_http_hosts,
+                global_secret_replacements, host_allowlist_snippet,
+                report_missing_outbound_http_secret_replacements,
+            },
+            create_engines, deployment_verify_config, fix_server_exec_digests, prepare_dirs,
             webhook_global_http_allowlist,
         },
         config::{
@@ -5571,7 +5212,6 @@ mod tests {
         let err = report_missing_outbound_http_secret_replacements(
             &missing,
             RuntimeConfigAvailability::Strict,
-            &crate::config::toml::ConfigWarnings::default(),
         )
         .unwrap_err()
         .to_string();
@@ -5610,7 +5250,6 @@ mod tests {
             &unregistered,
             "the deployment's outbound HTTP entries",
             RuntimeConfigAvailability::Strict,
-            &crate::config::toml::ConfigWarnings::default(),
         )
         .unwrap_err()
         .to_string();
@@ -5633,7 +5272,6 @@ mod tests {
             &unregistered,
             "the deployment's outbound HTTP entries",
             RuntimeConfigAvailability::AllowUnavailable,
-            &crate::config::toml::ConfigWarnings::default(),
         )
         .expect("unavailable runtime config downgrades unregistered secrets to a warning");
     }
@@ -5707,7 +5345,6 @@ mod tests {
         let err = report_missing_outbound_http_secret_replacements(
             &missing,
             RuntimeConfigAvailability::Strict,
-            &crate::config::toml::ConfigWarnings::default(),
         )
         .unwrap_err()
         .to_string();
@@ -5735,15 +5372,9 @@ mod tests {
         };
         let registry =
             SecretRegistry::from_test_values(Vec::<(String, secrecy::SecretString)>::new());
-        let warnings = crate::config::toml::ConfigWarnings::default();
         let global = GlobalHttpConfig::from(
-            crate::config::toml::resolve_allowed_hosts(
-                vec![host("obeli.sk")],
-                false,
-                &registry,
-                &warnings,
-            )
-            .unwrap(),
+            crate::config::toml::resolve_allowed_hosts(vec![host("obeli.sk")], false, &registry)
+                .unwrap(),
         );
         let name = crate::config::toml::ConfigName::new("caller".into()).unwrap();
 
@@ -5755,7 +5386,6 @@ mod tests {
             &global,
             &registry,
             false,
-            &warnings,
             &mut uncovered,
         );
 
@@ -5912,7 +5542,6 @@ mod tests {
             webui_enabled,
             server_verified.secret_registry,
             server_verified.global_http_config,
-            server_verified.config_warnings,
         ))
         .await?;
